initial migration from Primer::AutoComplete to Primer::Alpha::AutoComplete
