+++ previews/primer/alpha/tooltip_preview.rb
@@ -85,7 +85,7 @@ module Primer
 
       # @label Tooltip with Primer::IconButton
       def tooltip_with_icon_button(direction: :s, tooltip_text: "Tooltip text")
-        render(Primer::IconButton.new(icon: :search, "aria-label": tooltip_text, tooltip_direction: direction))
+        render(Primer::Beta::IconButton.new(icon: :search, "aria-label": tooltip_text, tooltip_direction: direction))
       end
       # @!endgroup
     end
