+++ static/constants.json
@@ -162,6 +162,8 @@
   },
   "Primer::Alpha::Image": {
   },
+  "Primer::Alpha::ImageCrop": {
+  },
   "Primer::Alpha::Layout": {
     "FIRST_IN_SOURCE_DEFAULT": "sidebar",
     "FIRST_IN_SOURCE_OPTIONS": [
